Add redirect_to_index_html boolean to [client]
It True, when web api has to display a directory node,
it first looks whether that directory contains an index.html file.
If so - it redirects to it. If not - it shows the folder view.
Useful for public gateways: some folders you want browsable,
but if you have a blog with "draft" posts, index.html protects
the folder from browsing.