"Decaf version" :) No JS, only folder level
Show permalinks as <input/> tags. I.e. JS can be disabled without breaking functionality.

To avoid clutter, there's only a folder level permalink, and the per-file ones are gone.
We can put them back if/when we redo the wapi in a way that makes it easy to say:
"if some_condition: drop the entire table column" (instead of the traditional "-" kludge).
Screenshot:
https://dubiousdod.org/uri/URI%3ADIR2-RO%3Asqvgvsl6pybotianh5sl7oo774%3Aqzhij6m4vmesi5reo6rgvbo62vdpaqqhts4dwa7bnede76tev7sq/no-js-permalink.png