.gitignore: ignore the generated zope.interface egg too
Some dependencies are installed in the top-level directory, maybe
because they're required by the build process itself? I'm not sure.
Twisted is one of these, if the system version is not new enough, and
our .gitignore has a clause to ignore the resulting .egg directory. Now
that we've bumped our zope.interface dependency, it sometimes builds a
z.i egg too (even if it doesn't build a Twisted one). So ignore both.