add a test, add missing imports
This tests ftpd, but not sftpd. Doing this sort of test on sftpd
requires the creation of a valid pubkey/privkey file pair, which is more
work than I want to do right now.

init_ftp/init_sftp were changed to interpret the configured
accounts.file as relative to the node's basedir, with
os.path.join(self.basedir, fn). This would happen naturally in a real
node, since it os.chdir()s to the basedir before doing anything. But
tests don't do that. Note that this os.path.join can mask the problem
which originally prompted this test,, because self.basedir is unicode,
and os.path.join(unicode,str) always gives you unicode.