Fix args passed to 'tahoe daemonize' when frozen
The code which now calls 'tahoe daemonize' in place of 'tahoe start'
always prepends sys.executable to the subsequent subprocess args list.
In the case of frozen builds, however, sys.executable points to the path
of the standalone frozen application file (or, in other words, frozen
python applications generally do not distinguish between a "python
interpreter" and a "python script" at runtime, the two being one and the
same file). The result is that the code in question effectively prepends
the frozen application path to the args list twice, thus calling 'tahoe
tahoe daemonize' (and failing) instead of properly calling 'tahoe
daemonize'. This change corrects this behavior by prepending
sys.executable to the args list only if the application is not running
in a 'frozen' state.