Always append setuptools to package_imports list
_auto_deps.py currently only appends setuptools to the package_imports
list if the application is running in a non-'frozen' state (citing that
it "triggers 'site' processing that causes failures"). Recent versions
of setuptools and/or PyInstaller, however, no longer appear to cause
such failures and, in the case of PyInstaller, frozen builds will
actually fail to run (with "PackagingError: no version info for
setuptools") if setuptools remains a dependency (i.e., is listed in
install_requires) but is *not* included in the package_imports list.
This change removes this no-longer-necessary conditional and appends
setuptools to package_imports always, thereby removing the need for the
unwieldy workaround currentnly present in the PyInstaller ".spec" file
(which comments out the 'setuptools' requirement at build time to avoid
the error at runtime) while still allowing frozen builds to execute
normally and without issue.