Remove test_client_no_noise and rewrite test_introducer
"tahoe run" has no quiet option so `test_client_no_noise` is not applicable.

This is a loss of the coverage of the quiet option for `tahoe start`.  That is
unfortunate but fixing any `tahoe start`-using test is really hard and the
functionality that is no longer covered is so trivial it hardly seems like it
made sense to test it by running multiple tahoe child processes anyway.